refactor: 계층형 구조로 패키지 세분화
