+++ src/main/java/com/abab/lectureRegister/exception/custom/LectureFullException.java
@@ -1,4 +1,4 @@
-package com.abab.lectureRegister.exception;
+package com.abab.lectureRegister.exception.custom;
 
 public class LectureFullException extends RuntimeException {
     public LectureFullException(Long lectureId) { super("Lecture ID " + lectureId + " is full");}
