+++ src/main/java/com/abab/lectureRegister/service/LectureService.java
@@ -1,7 +1,9 @@
-package com.abab.lectureRegister.lecture;
+package com.abab.lectureRegister.service;
 
-import com.abab.lectureRegister.exception.LectureFullException;
-import com.abab.lectureRegister.exception.LectureNotFoundException;
+import com.abab.lectureRegister.exception.custom.LectureFullException;
+import com.abab.lectureRegister.exception.custom.LectureNotFoundException;
+import com.abab.lectureRegister.model.Lecture;
+import com.abab.lectureRegister.repository.LectureRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
