+++ src/test/java/com/abab/lectureRegister/lecture/LectureServiceTest.java
@@ -1,7 +1,10 @@
 package com.abab.lectureRegister.lecture;
 
-import com.abab.lectureRegister.exception.LectureFullException;
-import com.abab.lectureRegister.exception.LectureNotFoundException;
+import com.abab.lectureRegister.exception.custom.LectureFullException;
+import com.abab.lectureRegister.exception.custom.LectureNotFoundException;
+import com.abab.lectureRegister.model.Lecture;
+import com.abab.lectureRegister.repository.LectureRepository;
+import com.abab.lectureRegister.service.LectureService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
