+++ src/test/java/com/abab/lectureRegister/registration/RegistrationTest.java
@@ -1,7 +1,10 @@
 package com.abab.lectureRegister.registration;
 
-import com.abab.lectureRegister.exception.LectureAlreadyRegisteredException;
-import com.abab.lectureRegister.lecture.Lecture;
+import com.abab.lectureRegister.exception.custom.LectureAlreadyRegisteredException;
+import com.abab.lectureRegister.model.Lecture;
+import com.abab.lectureRegister.model.Registration;
+import com.abab.lectureRegister.repository.RegistrationRepository;
+import com.abab.lectureRegister.service.RegistrationService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
